+++ src/matchers/element/toBeDisabled.ts
@@ -1,6 +1,11 @@
 import { executeCommandBe } from '../../utils'
+import { runExpect } from '../../util/expectAdapter'
 
 export function toBeDisabled(received: WebdriverIO.Element | WebdriverIO.ElementArray, options: ExpectWebdriverIO.CommandOptions = {}) {
+    return runExpect.call(this, toBeDisabledFn, arguments)
+}
+
+function toBeDisabledFn(received: WebdriverIO.Element | WebdriverIO.ElementArray, options: ExpectWebdriverIO.CommandOptions = {}) {
     this.expectation = this.expectation || 'disabled'
 
     return browser.call(async () => {
