+++ src/matchers/element/toBeDisplayed.ts
@@ -1,6 +1,11 @@
 import { executeCommandBe } from '../../utils'
+import { runExpect, getContext } from '../../util/expectAdapter'
 
 export function toBeDisplayed(received: WebdriverIO.Element | WebdriverIO.ElementArray, options: ExpectWebdriverIO.CommandOptions = {}) {
+    return runExpect.call(this, toBeDisplayedFn, arguments)
+}
+
+function toBeDisplayedFn(received: WebdriverIO.Element | WebdriverIO.ElementArray, options: ExpectWebdriverIO.CommandOptions = {}) {
     this.expectation = this.expectation || 'displayed'
 
     return browser.call(async () => {
@@ -16,6 +21,7 @@ export function toBeDisplayed(received: WebdriverIO.Element | WebdriverIO.Elemen
 }
 
 export function toBeVisible (el: WebdriverIO.Element, options: ExpectWebdriverIO.CommandOptions) {
-    this.expectation = 'visible'
-    return toBeDisplayed.call(this, el, options)
+    const context = getContext(this)
+    context.expectation = 'visible'
+    return toBeDisplayed.call(context, el, options)
 }
