+++ src/matchers/element/toBeSelected.ts
@@ -1,6 +1,11 @@
 import { executeCommandBe } from '../../utils'
+import { runExpect, getContext } from '../../util/expectAdapter'
 
 export function toBeSelected(received: WebdriverIO.Element | WebdriverIO.ElementArray, options: ExpectWebdriverIO.CommandOptions = {}) {
+    return runExpect.call(this, toBeSelectedFn, arguments)
+}
+
+function toBeSelectedFn(received: WebdriverIO.Element | WebdriverIO.ElementArray, options: ExpectWebdriverIO.CommandOptions = {}) {
     this.expectation = this.expectation || 'selected'
 
     return browser.call(async () => {
@@ -8,7 +13,9 @@ export function toBeSelected(received: WebdriverIO.Element | WebdriverIO.Element
         return result
     })
 }
+
 export function toBeChecked (el: WebdriverIO.Element, options: ExpectWebdriverIO.CommandOptions) {
-    this.expectation = 'checked'
-    return toBeSelected.call(this, el, options)
+    const context = getContext(this)
+    context.expectation = 'checked'
+    return toBeSelected.call(context, el, options)
 }
