+++ src/matchers/element/toHaveAttribute.ts
@@ -1,6 +1,11 @@
-import { waitUntil, enhanceError, compareText, executeCommand, getExpected, updateElementsArray } from '../../utils'
+import { waitUntil, enhanceError, compareText, executeCommand, wrapExpectedWithArray, updateElementsArray } from '../../utils'
+import { runExpect } from '../../util/expectAdapter'
 
 export function toHaveAttribute(received: WebdriverIO.Element | WebdriverIO.ElementArray, attribute: string, value?: string, options: ExpectWebdriverIO.StringOptions = {}) {
+    return runExpect.call(this, toHaveAttributeFn, arguments)
+}
+
+export function toHaveAttributeFn(received: WebdriverIO.Element | WebdriverIO.ElementArray, attribute: string, value?: string, options: ExpectWebdriverIO.StringOptions = {}) {
     const isNot = this.isNot
     const { expectation = 'attribute', verb = 'have' } = this
 
@@ -8,7 +13,7 @@ export function toHaveAttribute(received: WebdriverIO.Element | WebdriverIO.Elem
         let el = await received
         let attr
         const pass = await waitUntil(async () => {
-            const result = await executeCommand(el, condition, options, [attribute, value, options])
+            const result = await executeCommand.call(this, el, condition, options, [attribute, value, options])
             el = result.el
             attr = result.values
 
@@ -19,7 +24,7 @@ export function toHaveAttribute(received: WebdriverIO.Element | WebdriverIO.Elem
 
         let message: string
         if (typeof value === 'string') {
-            const expected = getExpected(el, attr, value)
+            const expected = wrapExpectedWithArray(el, attr, value)
             message = enhanceError(el, expected, attr, this, verb, expectation, attribute, options)
         } else {
             message = enhanceError(el, !isNot, pass, this, verb, expectation, attribute, options)
@@ -44,8 +49,5 @@ async function condition(el: WebdriverIO.Element, attribute: string, value: stri
         return { result: true, value: attr }
     }
 
-    return {
-        result: compareText(attr, value, options),
-        value: attr
-    }
+    return compareText(attr, value, options)
 }
