+++ src/matchers/element/toHaveAttributeContaining.ts
@@ -1,7 +1,12 @@
-import { toHaveAttribute } from './toHaveAttribute'
+import { runExpect } from '../../util/expectAdapter'
+import { toHaveAttributeFn } from './toHaveAttribute'
 
 export function toHaveAttributeContaining(el: WebdriverIO.Element, attribute: string, value: string, options: ExpectWebdriverIO.StringOptions = {}) {
-    return toHaveAttribute(el, attribute, value, {
+    return runExpect.call(this, toHaveAttributeContainingFn, arguments)
+}
+
+function toHaveAttributeContainingFn(el: WebdriverIO.Element, attribute: string, value: string, options: ExpectWebdriverIO.StringOptions = {}) {
+    return toHaveAttributeFn.call(this, el, attribute, value, {
         ...options,
         containing: true
     })
