+++ src/matchers/element/toHaveText.ts
@@ -1,6 +1,11 @@
-import { waitUntil, enhanceError, compareText, executeCommand, getExpected, updateElementsArray } from '../../utils'
+import { waitUntil, enhanceError, compareText, executeCommand, wrapExpectedWithArray, updateElementsArray } from '../../utils'
+import { runExpect } from '../../util/expectAdapter'
 
 export function toHaveText(received: WebdriverIO.Element | WebdriverIO.ElementArray, text: string, options: ExpectWebdriverIO.StringOptions = {}) {
+    return runExpect.call(this, toHaveTextFn, arguments)
+}
+
+export function toHaveTextFn(received: WebdriverIO.Element | WebdriverIO.ElementArray, text: string, options: ExpectWebdriverIO.StringOptions = {}) {
     const isNot = this.isNot
     const { expectation = 'text', verb = 'have' } = this
 
@@ -9,7 +14,7 @@ export function toHaveText(received: WebdriverIO.Element | WebdriverIO.ElementAr
         let actualText
 
         const pass = await waitUntil(async () => {
-            const result = await executeCommand(el, condition, options, [text, options])
+            const result = await executeCommand.call(this, el, condition, options, [text, options])
             el = result.el
             actualText = result.values
 
@@ -18,7 +23,7 @@ export function toHaveText(received: WebdriverIO.Element | WebdriverIO.ElementAr
 
         updateElementsArray(pass, received, el)
 
-        const message = enhanceError(el, getExpected(el, actualText, text), actualText, this, verb, expectation, '', options)
+        const message = enhanceError(el, wrapExpectedWithArray(el, actualText, text), actualText, this, verb, expectation, '', options)
 
         return {
             pass,
@@ -30,8 +35,5 @@ export function toHaveText(received: WebdriverIO.Element | WebdriverIO.ElementAr
 async function condition(el: WebdriverIO.Element, text: string, options: ExpectWebdriverIO.StringOptions) {
     let actualText = await el.getText()
 
-    return {
-        result: compareText(actualText, text, options),
-        value: actualText
-    }
+    return compareText(actualText, text, options)
 }
