+++ src/matchers/element/toHaveTextContaining.ts
@@ -1,7 +1,12 @@
-import { toHaveText } from './toHaveText'
+import { runExpect } from '../../util/expectAdapter'
+import { toHaveTextFn } from './toHaveText'
 
 export function toHaveTextContaining(el: WebdriverIO.Element, text: string, options: ExpectWebdriverIO.StringOptions = {}) {
-    return toHaveText(el, text, {
+    return runExpect.call(this, toHaveTextContainingFn, arguments)
+}
+
+function toHaveTextContainingFn(el: WebdriverIO.Element, text: string, options: ExpectWebdriverIO.StringOptions = {}) {
+    return toHaveTextFn.call(this, el, text, {
         ...options,
         containing: true
     })
