+++ src/matchers/element/toHaveValueContaining.ts
@@ -1,7 +1,12 @@
-import { toHaveValue } from './toHaveValue'
+import { runExpect } from '../../util/expectAdapter'
+import { toHaveValueFn } from './toHaveValue'
 
 export function toHaveValueContaining(el: WebdriverIO.Element, value: string, options: ExpectWebdriverIO.StringOptions = {}) {
-    return toHaveValue(el, value, {
+    return runExpect.call(this, toHaveValueContainingFn, arguments)
+}
+
+function toHaveValueContainingFn(el: WebdriverIO.Element, value: string, options: ExpectWebdriverIO.StringOptions = {}) {
+    return toHaveValueFn.call(this, el, value, {
         ...options,
         containing: true
     })
